requirements/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cos-python-sdk-v5==1.9.23
crcmod==1.7
cryptography==40.0.2
cssselect==1.1.0
cssutils==2.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-debug-toolbar==4.0.0
django-debug-toolbar-request-history==0.1.4
django-extensions==3.2.1
django-filter==23.1
django-import-export==3.2.0
django-watchman==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==4.8.3
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
premailer==3.10.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyMySQL==1.0.3
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
selenium==4.14.0
webdriver_manager==4.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
yagmail==0.15.293
zipp==3.6.0

uwsgi==2.0.21

# 爬虫
PyExecJS
aiohttp
drf-yasg==1.21.7

--index-url https://pypi.tuna.tsinghua.edu.cn/simple/